test: fix PSR-4 issues: naming, one-class-per-file
Closes #166
