+++ tests/Feature/Services/ApiRouteTest.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Services;
 
-use App\Providers\RouteServiceProvider;
 use App\Services\ApiRoute;
 use Illuminate\Routing\Router;
 use Illuminate\Support\Str;
@@ -176,31 +175,4 @@ class ApiRouteTest extends TestCase
         );
     }
 
-}
-
-class MockApiRouteServiceProvider extends RouteServiceProvider
-{
-    public function boot()
-    {
-        parent::boot();
-    }
-
-    public function map()
-    {
-        $this->apiRoute('test', 'TestController')
-            ->addDefaults()
-            ->add('/whatever', 'whatever')
-            ->add('/another', 'another', 'post');
-
-        $this->apiRoute('except', 'ExceptController')
-            ->except(['update']);
-
-        $this->apiRoute('only', 'OnlyController')
-            ->only(['index']);
-    }
-
-    public function apiRoute($prefix, $controller)
-    {
-        return new ApiRoute($prefix, $controller);
-    }
 }
